pyproject.toml
[tool.poetry]
name = "bundestag"
version = "0.1.0"
description = ""
authors = ["Andreas Domberg <andom@posteo.de>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "^1.26.4"
requests = "^2.32.2"
pyarrow = "^16.1.0"
plotly = "^5.22.0"
dash-bootstrap-templates = "^1.2.0"
nbformat = "^5.10.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
